[MAT-252] Rename the fake jni_md.h which clashed with the real one.
